requirements.txt
aenum==2.0.8
appdirs==1.4.3
attrs==18.1.0
black==18.6b2
click==6.7
GeoAlchemy==0.7.2
pg8000==1.12.1
psycopg2==2.7.4
python-dotenv==0.8.2
six==1.11.0
SQLAlchemy==1.2.8
sqlparse==0.2.4
testing.common.database==2.0.3
testing.postgresql==1.3.0
toml==0.9.4
